Add canonical badge match facts module
