+++ jupr_app/domain/gamification/evaluators.py
@@ -8,8 +8,8 @@ from typing import Any
 import pandas as pd
 
 from jupr_app.domain.awards import compute_top_performer_awards
-from jupr_app.domain.gamification.badge_rules import build_player_match_facts
 from jupr_app.domain.gamification.badge_types import BadgeCandidate, BadgeEvaluationContext
+from jupr_app.domain.gamification.match_facts import build_player_match_facts
 from jupr_app.domain.gamification.participation import compute_lifetime_games
 from jupr_app.domain.gamification.top_performer_awards import (
     TOP_PERFORMER_BADGE_IDS,
